Feature/introduction sprint guide (#49)
* feat: 🎸 add sprint guide page

* feat: 🎸 add layout

* feat: 🎸 add layout component

* feat: 🎸 add link and remove component
