+++ src/constants/introductionScrum.js
@@ -1,0 +1,25 @@
+export const scrumInfo = {
+  guide:
+    'Hi !!我是你的導覽員艾啾(agile)，想必你很想知道Scrum在做什麼吧！放心！我會帶你認識Scrum中的成員以及流程。',
+  titles: [
+    {
+      id: 0,
+      content: '首先，',
+    },
+    {
+      id: 1,
+      content: 'Scrum 是什麼？',
+    },
+  ],
+  sectionTitle: '基本概述',
+  image: {
+    type: 'animation',
+  },
+  article: [
+    'Scrum是一種敏捷開發的框架，以固定時間的長度週期持續開發產品，特色是以稱為固定時間長度的短衝（Sprint）週期去持續開發產品，並在每個短衝都為產品交付增量（Increment）—新增有價值的改變。',
+    '首先我們先來介紹scrum中的三大核心角色 產品負責人PO(productOwner)、Scrum主持人SM(Scrum Master)、開發團隊DEV(Developmnet Team)',
+  ],
+  button: '來看他們自我介紹',
+};
+
+export default scrumInfo;
