+++ src/views/introduction/SprintGuide.jsx
@@ -1,0 +1,12 @@
+import {
+  IntroductionLayout,
+} from '../../components/introduction/IntroductionLayout';
+import { sprintGuideInfo } from '../../constants/introductionSprintGuide';
+
+export function SprintGuide() {
+  return (
+    <IntroductionLayout info={sprintGuideInfo} />
+  );
+}
+
+export default SprintGuide;
