+++ src/views/introduction/index.js
@@ -1,2 +1,3 @@
 export * from './Scrum';
 export * from './ProductBacklog';
+export * from './SprintGuide';
