+++ tailwind.config.js
@@ -3,9 +3,6 @@ module.exports = {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
   theme: {
     extend: {
-      colors: {
-        primary: '#14B8A6',
-      },
       boxShadow: {
         'inner-blur-5px': 'inset 0px 0px 5px 1px rgb(0 0 0 / 0.05)',
       },
